[5.13] Updated bfq, arch-patches, lru, v4l2loopback, xanmod-patches
Also, kernels are now sorted from newest to oldest for easier editing/maintaining, so 5.13 will now be at the top, then 5.10, then 5.4.

Temporarily added an UBUNTU_PATCHES variable, which defaults to yes, for making the Ubuntu patches optional and making this less Ubuntu-dependent and more Debian friendly. Experimental for now but starting the ground work