Added latest CacULE scheduler patches for 5.10 and 5.13 (w/ RDB support)
The maintainer has merged the CacULE and RDB (an experimental load balancer based on interactivity scores) into one single patch now, which is nice, and the config option is turned off by default (CONFIG_CACULE_RDB). 

RDB is totally experimental, haven't even tried it myself. So use with caution, but at least it's easy to flip on now.

Otherwise, the standard CacULE has been working well so I'm just going to keep the defaults to the script the same to make it an easy path for beginners. Going to give RDB a try soon and see what it's all about, and will report back.