[5.14] Apply the cfs zen tweaks when using CacULE scheduler also
These tweaks should have been applied before but I was being lazy and didn't want to figure out the merge conflicts but it's been fixed now.

Since CacULE is built on top of CFS, we can leverage those base cfs zen tweaks and apply the CacULE specific ones that override it on top.

Also added a new file: zen-tweaks-cacule.patch, and it uses the KERNEL_SCHEDULER variable to apply zen-tweaks-cfs.patch or zen-tweaks-cacule.patch when the script is running depending on what variable is passed in.