[5.15] arch-patches-v3-sep -> arch-patches-v4-sep
Note: v3 -> v4 but no patch added as the patch that was added is already included in 5.15.

Please also note: patches are from the 5.15 lucjan branch, but copied over from 5.14 since they are still applicable.

For the time being while 5.15 is in release candidate, the patches for 5.14 and 5.15 will be sync'd in the build script, with the one ones that are already merged (or have a merge conflict) in 5.15 not included, at least not until the final release of 5.15 is released to avoid duplicating work.