Use the default CFS kernel scheduler by default instead of CacULE
It looks like the maintainer has archived the CacULE project on GitHub. So until further notice on the status of the project, the script will use CFS as the default kernel scheduler.

I think in general, this should be the new default behavior, since it's easy to flip on (just pass KERNEL_SCHEDULER=cacule to the script), and since CacULE is experimental and I want users to have the most stable experience, as opposited to maybe a slightly faster (but possibly) buggy experience. So far, CacULE has been great and I plan on keep using it myself.

Originally, I wanted more people to try out the CacULE scheduler without having to worry about custom variables, but moving forward, I think the default CFS just makes the most sense. We'll see.