Also, some prep work for the new kernel releases on 2021-09-26:
1) 5.13.19 no longer needs 0010-block-return-ELEVATOR_DISCARD_MERGE-if-possible.patch so it's been removed. Note:  5.13 is officialy end of life (EOL), so it shouldn't really be used anymore.

2) 5.14.2-rt and higher now have a merge conflict on miscellaneous fix: 0015-atlantic-Fix-issue-in-the-pm-resume-flow.patch

Since it's still in -rc1, we'll just remove this patch, but since it's still good for non-rt (regular builds), we'll leave it for those.

3) 5.10.69 and higher now have a merge conflict on 0019-block-bfq-merge-burst-of-newly-created-queues.patch

Instead of just removing it, let's add a merge conflict fix (worth it, since 5.10 is LTS and supported until 2026). If we had just removed this patch, it may have still result in everything to patch cleanly, and even compile fine, but may have had unintended consequences during use. Always the better approach when handling merge conflicts/issues in case you run into something similar.