[5.13 / 5.14 / 5.15] Fix build issues on a new install
So on a fresh Ubuntu install, I noticed the line for patching cfs-zen-tweaks was looking for the wrong file which didn't exist, so the builds have been failing for 5.13, 5.14, and 5.15 as of the past couple weeks, but not the main builds at least. Big oof, my bad.

Will be also adding a readme soon with the pre-prerequisites to get going on a new installation so you can just run ./build_kernel.sh and be good to go.

Lastly, the io_uring block patches by Jens Axboe are no longer patching cleanly in the latest 5.15-rc, so we'll keep the patches in the tree, but remove them from being applied in the script. These patches were only two of a much larger set, so it's probably best to leave them out and wait for them to be fully merged in the next 5.16 kernel, rather than cherry picking one or two.