Updated compatibility with lucjan-patches repository
Good news: lucjan has cleaned up the lucjan-patches repository and removed all 4.x kernel versions and all 5.x versions before 5.10. 

This is nice as we had to pull in the entire repository even though we weren't using most of them. (We appreciate him maintaining these patches, as this script heavily relies on them.)

As a result, I've added the 5.4 through 5.9 patches to this repository (but only the ones that were being used by the build script for building 5.4) and fixed/updated the script to properly build the 5.4 and 5.4-rt kernels. (important, as it is still the default kernel being built when you run ./build_kernel.sh).

readme coming soon, promise