[5.15] Add damon-patches
Disabled by default. See kernel config to enable.

DAMON is a Linux kernel module made by Amazon for monitoring of data accesses for specific user-space processes.

Here is how Amazon's team explains DAMON (via Phoronix):

DAMON is a kernel module that allows users to monitor the actual memory access pattern of specific user-space processes. It aims to be 1) accurate enough to be useful for performance-centric domains, and 2) sufficiently light-weight so that it can be applied online.

For the goals, DAMON utilizes its two core mechanisms, called region-based sampling and adaptive regions adjustment. The region-based sampling allows users to make their own trade-off between the quality and the overhead of the monitoring and set the upper-bound of the monitoring overhead. Further, the adaptive regions adjustment mechanism makes DAMON to maximize the quality and minimize the overhead with its best efforts while preserving the users configured trade-off.