[5.15 / 5.16] cpufreq-patches-v4 -> cpufreq-patches-v5
Note: These patches are the new AMD pstate driver for Zen 2, 3, and future processors.

Also note: v5 of the patches = v6 of the driver, so these are the latest available.

The amd_pstate driver works best with the performance governor instead of schedutil.

The driver can be enabled in GRUB by adding the following to GRUB_CMDLINE_LINUX_DEFAULT:

amd_pstate.enable=1 amd_pstate.shared_mem=1