[5.15 / 5.16] cpufreq-patches-v7 -> cpufreq-patches-v8
These patches are the new AMD p-state driver for Zen 2, 3, and future processors.

v7 of the patches = v8 of the driver, so these are the latest available.

The amd-pstate driver works best with the performance governor instead of schedutil.

The driver can be enabled in GRUB by adding the following to GRUB_CMDLINE_LINUX_DEFAULT:

amd_pstate.enable=1 amd_pstate.shared_mem=1