Fix ubuntu-defaults config for 5.16, 5.17, and 5.18
For the past three versions (5.16, 5.17, 5.18) they were incorrectly duplicates of the custom config used in this script and not reflecting the Ubuntu defaults. 

To recap: the ubuntu-defaults configs are used as a reference to see the changes of the custom config so you can see what's different, or to use the Ubuntu base config as a starting point instead.