Fix deprecated dependency of crda/wireless-crda from kernel builds
More info:

Since the release of Ubuntu 22.04, the crda and wireless-crda packages have been deprecated and no longer available from Ubuntu and Debian upstream. These were previously used as a way to set the regulatory domain for Wi-Fi.

This would result in a broken package (see error message below) on the system for the kernel modules package created by the build script, as well as broken Wi-Fi if you resolved the broken package (via apt --fix-broken) and then rebooted into the new kernel.

The following packages have unmet dependencies:
 linux-modules-extra-6.9.5-060905+customidle-generic : Depends: crda but it is not installable or
                                                                wireless-crda but it is not installable
E: Unmet dependencies. Try 'apt --fix-broken install' with no packages (or specify a solution).

This never happened on my machine since I was on 21.10 originally and upgraded to 22.04, so the crda / wireless-crda were already present on my machine. However, on a fresh install of 22.04 and upwards, the kernel build script would ultimately fail at the end with the above message.

This commit removes the no longer needed "depends on" for these packages in the Ubuntu sauce patches and allows the kernel modules to successfully build.

In case the change doesn't show in the changelog due to file size, it consists of updaing the following line:

+Depends: ${misc:Depends}, ${shlibs:Depends}, linux-image-PKGVER-ABINUM-FLAVOUR | linux-image-unsigned-PKGVER-ABINUM-FLAVOUR, crda | wireless-crda

to

+Depends: ${misc:Depends}, ${shlibs:Depends}, linux-image-PKGVER-ABINUM-FLAVOUR | linux-image-unsigned-PKGVER-ABINUM-FLAVOUR

More info below:

https://bugs.launchpad.net/network-manager/+bug/1951586
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1958918