Remove support for AMDGPU_BUILTIN
I recently upgraded from Ubuntu 22.04 to Ubuntu 24.04 and the files in /lib/firmware/amdgpu are now compressed (.zst for zstd). To fix this, we can enable CONFIG_FW_LOADER_COMPRESS_ZSTD in the config, but if someone has a different compression (LZ4, for example), that won't work. So I'm just removing this altogether. Although there are performance improvements to baking in the amdgpu module into the kernel, it's not worth the upkeep anymore. Not to mention, ROCM requires amdgpu to be a module for it to work, so let's get rid of it.